+++ agent_atlas/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 """Agent Atlas package."""
 
-__version__ = "0.1.4"
+__version__ = "0.1.5"